Add dockerfile
This dockerfile is based on the officially available Nix image, and therefore contains the nix package manager and git by default. Vim was manually installed, as to allow a more complete experience in the event quick file editing is necessary. When someone opens a container, they are already in the repository's working directory. Lastly, to solve an issue pertaining to haskeline, cabal update was called; if the reliance on this command is deemed as problematic in future, it is recommended to upload the image generated by this file to a hosting service, such as Github Packages or Dockerhub, as to avoid multiple installations at different points in time.